modulo_3-Construção-e-deply-de-agentes-de-ia/1_agentes-ia-e-eprocesso-de-automacao/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==7.3.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
ffmpy==0.5.0
filelock==3.17.0
frozendict==2.4.6
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.14.0
gradio_client==1.7.0
groq==0.16.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.24.1
numpy==2.2.2
openai==1.60.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peewee==3.17.8
phidata==2.7.10
pillow==11.1.0
platformdirs==4.3.6
primp==0.11.0
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
ruff==0.9.4
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
wheel==0.45.1
yfinance==0.2.52
